+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [NEXT]
+### Improvements
+- do not try to disable window compositors for wayland sessions
+
 ## [1.3.3] 2023-08-12
 
 ### Improvements